Reduce TODO count to under 100
Prevent possible OOB write for label search
Confirmed ATAN 2-arg form has no errors based on numerical arguments in PTC
Moved variable functions out from system.h into variables.h
Softer failure for invalid wide to small character conversion
str_comp now doesn't use a bunch of stack
Confirmed BGMPLAY maxes out at 9 arguments in PTC
INPUT now uses COLOR values for user text + test
Confirmed GPAINT offscreen is valid in PTC
Fixed lowest bit of green color being lost in COLSET/COLREAD + tests

+++ source/interpreter/strs.h
@@ -88,11 +88,14 @@ bool is_varname(char c);
 u32 name_hash(const char* name, const u32 len, const u32 hmax);
 bool namecmp(const char* a, const u32 len, const char b[16]);
 
-//Convert char to wide char
+// Convert char to wide char
 u16 to_wide(u8 c);
 
-//Reverse char to wide char
-u8 to_char(u16 w);
+/// Converts a wide character into an 8-bit character
+///
+/// @param w Wide character to convert
+/// @return 8-bit character code, or -1 if wide character is out of range
+int to_char(u16 w);
 
 //Convert a (hex) digit into an integer;
 int digit_value(const u16 c);
